modificaciones BBDD nuevos metodos en DAOS para acceso a BBDD

+++ administracion/administracion-model/src/main/java/es/uc3m/tiw/daos/MaterialLeccionDao.java
@@ -12,6 +12,6 @@ public interface MaterialLeccionDao {
 
 	public abstract List<MaterialLeccion> findAll() throws Exception;
 	
-	public abstract MaterialLeccion findById(int idMaterial) throws Exception;
+	public abstract MaterialLeccion findById(Long idMaterial) throws Exception;
 
 }
