+++ administracion/administracion-model/src/main/java/es/uc3m/tiw/daos/OfertaDao.java
@@ -12,5 +12,5 @@ public interface OfertaDao {
 
 	public abstract List<Oferta> findAll() throws Exception;
 	
-	public abstract Oferta findById(int idOferta) throws Exception;
+	public abstract Oferta findById(Long idOferta) throws Exception;
 }
